requirements.txt
setuptools~=75.8.0
opencv-python~=4.11.0.86
pydicom~=3.0.1
numpy~=2.2.2
pynrrd~=1.1.3
